backend/requirements.txt
setuptools==69.5.1
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
django-vite==3.1.0
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.2.2
pip==25.0.1
pipreqs==0.4.13
psycopg-binary==3.2.6
python-decouple>=3.8
psycopg2-binary>=2.9.10
